3. fix menu commands c. Log command when active should have check box.
Added a closeLogWindow() method to the LogWindow Class
When the log command is executed, the checkbox is selected and the log window opens. 
When the log command is clicked again, the checkbos is deselected and the log window closes.
However, when the window is manually closed the checkbox does not change